+++ src/noahtheduke/splint/rules/lint/body_unquote_splicing.clj
@@ -10,15 +10,11 @@
 (set! *warn-on-reflection* true)
 
 (defn only-body [sexp]
-  (case sexp
-    (delay dosync future lazy-cat lazy-seq pvalues with-loading-context) true
-    false))
+  ('#{delay dosync future lazy-cat lazy-seq pvalues with-loading-context} sexp))
 
 (defn init-arg [sexp]
-  (case sexp
-    (binding locking sync with-bindings with-in-str with-local-vars
-      with-precision with-redefs) true
-    false))
+  ('#{binding locking sync with-bindings with-in-str with-local-vars
+      with-precision with-redefs} sexp))
 
 (defn only-body-new-form [?f ?body]
   (list ?f (list 'let ['res# (list 'do (list 'splint/unquote-splicing ?body))] 'res#)))
@@ -40,12 +36,11 @@
   ; good
   `(binding [max mymax] (let [res# (do ~@body)] res#))
   "
-  {:patterns ['((? f only-body) (splint/unquote-splicing ?body))
-              '((? f init-arg) ?init-arg (splint/unquote-splicing ?body))]
+  {:patterns ['((? f only-body) (splint/unquote-splicing (? body symbol?)))
+              '((? f init-arg) ?init-arg (splint/unquote-splicing (? body symbol?)))]
    :message "Wrap ~@/unquote-splicing in `(let [res# (do ...)] res#)` to avoid unhygenic macro expansion."
    :on-match (fn [ctx rule form {:syms [?f ?init-arg ?body]}]
-               (when (symbol? ?body)
-                 (let [new-form (if ?init-arg
-                                  (init-arg-new-form ?f ?init-arg ?body)
-                                  (only-body-new-form ?f ?body))]
-                   (->diagnostic ctx rule form {:replace-form new-form}))))})
+               (let [new-form (if ?init-arg
+                                (init-arg-new-form ?f ?init-arg ?body)
+                                (only-body-new-form ?f ?body))]
+                 (->diagnostic ctx rule form {:replace-form new-form})))})