+++ src/noahtheduke/splint/rules/style/useless_do.clj
@@ -10,6 +10,9 @@
 
 (set! *warn-on-reflection* true)
 
+(defn not-unquote-splicing [sexp]
+  (not (when (sequential? sexp) (unquote-splicing?? (first sexp)))))
+
 (defrule style/useless-do
   "A single item in a `do` is a no-op. However, it is sometimes necessary to wrap expressions in `do`s to avoid issues, so `do` surrounding `~@something` will be skipped as well as `#(do something)`.
 
@@ -25,11 +28,10 @@
   (do ~@body)
   #(do [%1 %2])
   "
-  {:pattern '(do ?x)
+  {:pattern '(do (? x not-unquote-splicing))
    :message "Unnecessary `do`."
    :on-match (fn [ctx rule form {:syms [?x]}]
-               (when-not (when (sequential? ?x) (unquote-splicing?? (first ?x)))
-                 (let [parent-form (:parent-form ctx)]
+               (let [parent-form (:parent-form ctx)]
                    (when-not (and (sequential? parent-form)
                                   (= 'splint/fn (first parent-form)))
-                     (->diagnostic ctx rule form {:replace-form ?x})))))})
+                     (->diagnostic ctx rule form {:replace-form ?x}))))})