+++ test/noahtheduke/splint/re_frame_test.clj
@@ -18,10 +18,9 @@
 
 (def re-frame-diagnostics
   '{lint/if-else-nil 2
-    lint/prefer-method-values 1
     lint/thread-macro-one-arg 15
     lint/try-splicing 1
-    lint/warn-on-reflection 4
+    lint/warn-on-reflection 6
     metrics/fn-length 20
     naming/conventional-aliases 2
     naming/predicate 1
@@ -51,4 +50,4 @@
              :diagnostics
              (group-by :rule-name)
              (#(update-vals % count)))))
-    (expect 62 (count (:diagnostics results)))))
+    (expect 63 (count (:diagnostics results)))))